Add integration tests for setting plain text values
