+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/config/set/__scripts__/set.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+FORCE_COLOR=0
+
+mkdir -p test
+cd test
+imperative-test-cli config set $1 $2 $3
+exit $?
