+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/config/set/__scripts__/set_secure.sh
@@ -4,5 +4,5 @@ FORCE_COLOR=0
 
 mkdir -p test
 cd test
-imperative-test-cli config set-secure $1 $2 $3
+imperative-test-cli config set $1 $2 $3 --secure
 exit $?
